change eeprom dbm to float
Change everything to use dbm where possible.

+++ src/src/openVTxEEPROM.h
@@ -21,7 +21,7 @@ typedef struct
     uint8_t freqMode;
     uint8_t pitmodeInRange;
     uint8_t pitmodeOutRange;
-    uint8_t currPowerdB;
+    float currPowerdB;
     uint8_t currPowermW; // Required due to rounding errors when converting between dBm and mW
     uint8_t unlocked;
 } openVTxEEPROM;
